Move Call.Result as upper level class
